error handling

+++ src/main/java/org/onstage/team/service/TeamService.java
@@ -2,11 +2,12 @@ package org.onstage.team.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.onstage.exceptions.BadRequestException;
 import org.onstage.team.client.TeamDTO;
 import org.onstage.team.model.Team;
 import org.onstage.team.repository.TeamRepository;
 import org.springframework.stereotype.Service;
+
+import static org.onstage.exceptions.BadRequestException.teamNotFound;
 
 @Service
 @Slf4j
@@ -15,23 +16,25 @@ public class TeamService {
     private final TeamRepository teamRepository;
 
     public Team getById(String id) {
-        Team team = teamRepository.getById(id);
-        if (team == null) {
-            throw BadRequestException.teamNotFound();
-        }
-        return team;
+        return teamRepository.getById(id);
     }
 
     public Team save(Team team) {
-        return teamRepository.save(team);
+        Team savedTeam = teamRepository.save(team);
+        log.info("Team {} has been saved", savedTeam.id());
+        return savedTeam;
     }
 
     public String delete(String id) {
+        if (teamRepository.getById(id) == null) {
+            throw teamNotFound();
+        }
+        log.info("Deleting team {}", id);
         return teamRepository.delete(id);
     }
 
-    public Team update(String id, TeamDTO request) {
-        Team existingTeam = getById(id);
+    public Team update(Team existingTeam, TeamDTO request) {
+        log.info("Updating team {} with request {}", existingTeam.id(), request);
         Team updatedTeam = updateTeamFromDTO(existingTeam, request);
         return save(updatedTeam);
     }
