+++ src/main/java/org/onstage/teammember/service/TeamMemberService.java
@@ -3,9 +3,6 @@ package org.onstage.teammember.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.onstage.exceptions.BadRequestException;
-import org.onstage.team.client.TeamDTO;
-import org.onstage.team.model.Team;
-import org.onstage.team.repository.TeamRepository;
 import org.onstage.teammember.model.TeamMember;
 import org.onstage.teammember.repository.TeamMemberRepository;
 import org.springframework.stereotype.Service;
@@ -17,18 +14,20 @@ public class TeamMemberService {
     private final TeamMemberRepository teamMemberRepository;
 
     public TeamMember getById(String id) {
-        TeamMember teamMember = teamMemberRepository.getById(id);
-        if (teamMember == null) {
-            throw BadRequestException.teamMemberNotFound();
-        }
-        return teamMember;
+        return teamMemberRepository.getById(id);
     }
 
     public TeamMember save(TeamMember teamMember) {
-        return teamMemberRepository.save(teamMember);
+        TeamMember savedTeamMember = teamMemberRepository.save(teamMember);
+        log.info("Team member {} has been saved", savedTeamMember.id());
+        return savedTeamMember;
     }
 
     public String delete(String id) {
+        if (teamMemberRepository.getById(id) == null) {
+            throw BadRequestException.teamMemberNotFound();
+        }
+        log.info("Deleting team member {}", id);
         return teamMemberRepository.delete(id);
     }
 }
